update double tap

+++ TP1Partie3/TP1Partie3/Competition.h
@@ -25,7 +25,7 @@
 -(void)ajouterAthlete:(NSString *)pPrenom: (NSString *)pNom: (NSString *)pPays;
 -(int)nouveauNumeroAthlete;
 -(Athlete *)getAthleteCourant;
--(void)setPointageAthleteCourant:(int)pPointageJuges:(double)pTemps;
+-(void)setPointageAthleteCourant:(double)pPointageJuges:(double)pTemps;
 -(NSMutableArray *)setPosition;
 -(NSArray *)getLstAthlete;
 -(NSArray *)getLstAthleteAttente;
